moved lines in lines folder
