+++ src/app/dashboard/index.component.ts
@@ -36,10 +36,10 @@ import { TableService } from '@services/table.service';
 import { TasksByStatusService } from '@services/tasks-by-status.service';
 import { UtilsService } from '@services/utils.service';
 import { AddLineDialogComponent } from './components/add-line-dialog.component';
-import { ApplicationsLineComponent } from './components/applications-line.component';
+import { ApplicationsLineComponent } from './components/lines/applications-line.component';
 import { ReorganizeLinesDialogComponent } from './components/reorganize-lines-dialog.component';
 import { SplitLinesDialogComponent } from './components/split-lines-dialog.component';
-import { LineComponent } from './components/task-status-line.component';
+import { LineComponent } from './components/lines/task-status-line.component';
 import { DashboardIndexService } from './services/dashboard-index.service';
 import { DashboardStorageService } from './services/dashboard-storage.service';
 import { Line } from './types';
